% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtm.R
\name{eqtm}
\alias{eqtm}
\title{eqtm}
\usage{
eqtm(dmrs_gr, gene_col, meth_data, expr_data, meth_anno_gr,
  cor_method = c("pearson", "kendall", "spearman"),
  alternative = c("two.sided", "greater", "less"), N = 1000, ncores = 1,
  iseed = NULL, verbose = T)
}
\arguments{
\item{dmrs_gr}{A GenomicRanges-derived object where the ranges represent the coordinates of the DMR and whose metadata contains an associated gene identifier.}

\item{gene_col}{A column index containing of the dmrs_gr object containing the gene identifiers.}

\item{meth_data}{A matrix whose row names contain CpG identifiers and whose values represent the methylation signal. The column names must contain the columns names of the gene expression data.}

\item{expr_data}{A matrix whose row names contain gene identifiers and whose values represent the expression signal. The column names must contain the columns names of the methylation data.}

\item{meth_anno_gr}{A GenomicRanges-derived object where the ranges represent the coordinates of the assayed CpGs and whose rownames contain an identifier.}

\item{alternative}{Alternative hypothesis to be used for calculating the p-values ("two.sided", "greater", "less").}

\item{N}{The number of bootstraps and permutations for calculating the 95% confidence intervals and p-values respectively (default: 1000).}

\item{ncores}{The number of cores to use for bootstrapping and permutations (default: 1)}

\item{iseed}{A seed for the bootstrapping and permutations for reproducibility}

\item{verbose}{A boolean whether the printed output should be verbose (default: T)}

\item{cor_type}{Type of correlation ("pearson", "kendall", "spearman").}
}
\value{
A SummarizedExperiment object containing the location of the summarized methylation and gene expression data for the overlapping samples.
}
\description{
This function was setup to perform an expression quantitative trait methylation analysis for methylation and gene expression data obtained from the same samples.
}
\author{
Andrew Y.F. Li Yim
}
\keyword{eqtm,}
\keyword{expression}
\keyword{methylation,}
