% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtm_plot.R
\name{plot_eqtm}
\alias{plot_eqtm}
\title{plot_eqtm}
\usage{
plot_eqtm(dmrs_se, meth_data, meth_groups, anno_gr, expr_data, expr_groups,
  united_groups, bm, index, colors = NULL)
}
\arguments{
\item{dmrs_se}{A SummarizedExperiment object generated from the eqtm() function}

\item{sort.by}{What to sort by? ("pval", "cor_coef", "CI95_diff")}

\item{volcanoplot}{A boolean whether or not to generate a volcano plot (x = correlation, y = -log10(p-value))}
}
\description{
This function was setup to perform an expression quantitative trait methylation analysis for methylation and gene expression data obtained from the same samples.
}
\author{
Andrew Y.F. Li Yim
}
\keyword{eqtm,}
\keyword{expression}
\keyword{methylation,}
