% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topcor.R
\name{topcor}
\alias{topcor}
\title{topcor}
\usage{
topcor(dmrs_se, sort.by = c("pval", "cor_coef", "CI95_diff"),
  volcanoplot = T)
}
\arguments{
\item{dmrs_se}{A SummarizedExperiment object generated from the eqtm() function}

\item{sort.by}{What to sort by? ("pval", "cor_coef", "CI95_diff")}

\item{volcanoplot}{A boolean whether or not to generate a volcano plot (x = correlation, y = -log10(p-value))}
}
\value{
A data.frame containing the eqtms ranked according to the 'sort.by' argument
}
\description{
This function extracts the
}
\author{
Andrew Y.F. Li Yim
}
\keyword{eqtm,}
\keyword{expression}
\keyword{methylation,}
